Switch to P2P
I'm not going to spend a lot of money on servers so P2P is a good solution.

With bevy_matchbox, you get easy P2P messaging as a Bevy
resource. MatchboxSocket can be used as a normal Res, whereas the WebSocket
solutions I was using had to be NonSend.

Because the connection integrates into the Bevy app, it can coexist with the
simulation code. With the server-client architecture, it would be very weird
spawning a Bevy app inside a Tokio app. That would be a thread pool inside
another thread pool.

Note: for some reason (my networking knowledge is lacking) the static server
needs to listen on 0.0.0.0 instead of localhost. So make it configurable and set
the default.

At the moment, you need to run a signaling server, e.g. matchbox_server:

cargo install matchbox_server
~/.cargo/bin/matchbox_server